+++ src/components/tips/Tips.aat.tsx
@@ -41,12 +41,12 @@ describe('Tips offers data and view-control', () => {
     })
   })
 
-  describe('in tough play-mode', () => {
-    const TOUGH_CONFIG = { ...microConfig, PLAY_MODE: PlayMode.TOUGH }
+  describe('in bare play-mode', () => {
+    const BARE_CONFIG = { ...microConfig, PLAY_MODE: PlayMode.BARE }
     let screen: RenderResult
 
     beforeEach(async () => {
-      screen = await renderWithContext(TipsCmp, { config: TOUGH_CONFIG })
+      screen = await renderWithContext(TipsCmp, { config: BARE_CONFIG })
     })
 
     it('should show clock', async () => {
@@ -62,7 +62,7 @@ describe('Tips offers data and view-control', () => {
     })
 
     it('should indicate the mode', async () => {
-      const indicator = screen.getByText(/Tough/i)
+      const indicator = screen.getByText(/Bare/i)
       expect(indicator).toBeInTheDocument()
     })
   })